requirements.txt
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.3
bjoern==3.2.2
build==1.0.3
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==42.0.2
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
dictop==0.2.2
Django==4.2.11
django-ajax-selects==3.0.2
django-anymail==10.2
django-appconf==1.0.6
django-baton==2.9.1
django-countries==7.5.1
django-dbbackup==4.1.0
django-debug-toolbar==4.3.0
django-defender==0.9.7
django-dramatiq==0.11.6
django-extra-views==0.14.0
django-filter==23.3
django-froala-editor==4.1.2
django-haystack==3.2.1
django-lifecycle==1.1.2
django-modeltranslation==0.18.11
django-oscar==3.2.4
django-oscar-api==3.2.4
django-phonenumber-field==6.4.0
django-redis==5.4.0
django-rest-knox==4.2.0
django-silk==5.1.0
django-static-fontawesome==6.4.2.1
django-static-jquery3==5.3.7.1
django-tables2==2.3.4
django-treebeard==4.7
django-user-agents==0.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
dramatiq==1.14.2
drf-spectacular==0.27.1
drf-yasg==1.21.7
easy-thumbnails==2.8.5
exceptiongroup==1.1.3
executing==2.0.0
factory-boy==3.2.1
Faker==19.6.2
google-auth==2.29.0
gprof2dot==2022.7.29
hiredis==2.3.2
idna==3.4
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
magic-import==0.2.2
markdown2==2.4.10
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.29
pickleshare==0.7.5
pillow==10.2.0
pip-tools==7.3.0
pipdeptree==2.13.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.17
psycopg-binary==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
purl==1.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycountry==23.12.11
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyquery==2.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
readtime==3.0.0
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rest_social_auth==8.3.0
rpds-py==0.10.4
rsa==4.9
sentry-sdk==1.40.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sorl-thumbnail==12.10.0
soupsieve==2.5
South==1.0.2
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.11.2
tus.py==1.3.4
typing_extensions==4.8.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.7
urlman==2.0.1
user-agents==2.2.0
wcwidth==0.2.8
